requirements.txt
pandas==2.2.2
sqlalchemy==2.0.30
psycopg2==2.9.9
python-dotenv==1.0.1